% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{averageEuclideanDistance}
\alias{averageEuclideanDistance}
\title{Average Euclidean Distance}
\usage{
averageEuclideanDistance(lambda, aggregationFactor = 1, epsilon = 0)
}
\arguments{
\item{lambda}{the average dialy movement distance of an individual (in
kilometers).}

\item{aggregationFactor}{the factor of aggregation applied to the raster
data mentioned in the function details (expressed in kilometers).}

\item{epsilon}{a rounding error correction (a real number; by default,
zero).}
}
\value{
a matrix of the average Euclidean distances
}
\description{
Measure the exposure influence upon susceptible individuals at
spatial position (x, y) of the infectious individuals at a spatial
position (u, v).
}
\details{
The 2020 United Nations-adjusted population count raster data, at
the 30-arc second resolution, consists of hundreds of thousands (and for
some countries millions) of grid cells. This function is used to calculate
the average Euclidean distance from one cell to others, adjusting for
raster aggregation.

The mathematical modelling of human or non-human mobility patterns is
non-trivial. This function is a limited implementation of the effective
area for only human mobility, with distances travelled per day (lambda)
measured in kilometers.

NOTE: Our weight function does not take the same arguments as shown in the
slideshow: (w x y u v); The term "effective area" comes from the fact that
if the kernel is constant on a finite disk (and zero outside it), then the
formula due to Bolker (1999) gives the area of the disk.

See the article titled \emph{A clarification of transmission terms in
host-microparasite models by Begon et al.} (2002).

The count of infected persons, or incidence, raster data provided to the
caller of this function, i.e. \link{transmissionLikelihoodWeightings}, may be
aggregated by a given factor greater than one, \code{aggregationFactor}; it is
passed regardless.
}
\examples{
averageEuclideanDistance(lambda = 15) # No raster aggregation
averageEuclideanDistance(lambda = 15, aggregationFactor = 35)

lattice::levelplot(as.array(averageEuclideanDistance(15)))
lattice::levelplot(as.array(averageEuclideanDistance(100, 35)))
}
\author{
Bryce Carson

Thomas White
}
\keyword{internal}
