% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{avgEuclideanDistance}
\alias{avgEuclideanDistance}
\title{Average Euclidean Distance}
\usage{
avgEuclideanDistance(radius, lambda, aggregationFactor = NULL)
}
\arguments{
\item{radius}{TODO: a fixed radius r > lambda; see details.}

\item{lambda}{movemenet distance (in kilometers) per day; see details.}

\item{aggregationFactor}{the degree of aggregation applied to the raster
data mentioned in the function details.}
}
\value{
a matrix of the average Euclidean distances
}
\description{
Measure the exposure influence upon susceptible individuals at
spatial position (x, y) of the infectious individuals at a spatial
position (u, v).
}
\details{
The mathematical modelling of human or non-human mobility patterns
is non-trivial. This function is a limited implementation of the effective
area for only human mobility, with distances travelled per day (lambda)
measured in kilometers.

NOTE: Our weight function does not take the same arguments as shown in the
slideshow: (w x y u v); The term "effective area" comes from the fact that
if the kernel is constant on a finite disk (and zero outside it), then the
formula due to Bolker (1999) gives the area of the disk.

See the article titled \emph{A clarification of transmission terms in
host-microparasite models by Begon et al.} (2002).

The raster data of infection counts or disease incidence provided to the
function which calls this one, transmissionLikelihoodWeightings, may be
aggregated by a given factor. That factor must be passed to this function
for parity, so the data is treated the same.
}
\examples{
avgEuclideanDistance(radius = 1, lambda = 15)
avgEuclideanDistance(radius = 1, lambda = 15, aggregationFactor = 35)
}
\author{
Bryce Carson

Thomas White
}
\keyword{internal}
