% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{classify.binary}
\alias{classify.binary}
\title{Create a binary reclassification of the input SpatRaster}
\usage{
classify.binary(inputRaster)
}
\arguments{
\item{inputRaster}{A single-layer SpatRaster}
}
\value{
a binary reclassification of the input SpatRaster
}
\description{
The susceptible matrix is reclassified so positive values become one and
negative values become zero. Values which are zero remain zero. This results
in a binary classification of inhabitance when given a SpatRaster with
population count or susceptible data; regardless, whatever input SpatRaster
is provided, a binary classification of the values in the raster are
returned in a SpatRaster of equal dimension and with the same coordinate
reference system.
}
\details{
The (re)-classification function built-in to terra is much faster than an
ad-hoc implementation of the same behaviour, of course, so the built-in
method is used (though it requries the reader to learn a little more of the
raster library; a good, a necessary, penalty).
}
\author{
Bryce Carson
}
