% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{.onLoad}
\alias{.onLoad}
\title{Cause side-effects when loading this package}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The name of the library loaded}

\item{pkgname}{The name of the package loaded}
}
\description{
Set the options for this package, primarily set path
\code{spatialEpisim.foundation.datapath} used for storing data used by this
package which is downloaded from the Web.
}
\details{
The option \code{spatialEpisim.foundation.datapath} is only set to its default
value if it is not already set.
}
\author{
Bryce Carson
}
