% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{downloadWorldPopData}
\alias{downloadWorldPopData}
\title{Download WorldPop 2020 data}
\usage{
downloadWorldPopData(countryISO3C, folder = here::here("data", "geotiff"))
}
\arguments{
\item{countryISO3C}{The uppercase ISO three character code a recognized
country.}

\item{folder}{The destination folder the downloaded data will be stored in}
}
\value{
An absolute path where the data was downloaded, or the path at which
the file already existed
}
\description{
Download 1 km aggregated UN-adjusted population count spatial data from 2020
through the WorldPop servers.
}
\details{
This function only downloads the file if necessary (if it doesn't exist at
path).

ppp is the code used by WorldPop to mean population estimate for population
counts; similar data, but which grids population density is coded \emph{pd}
rather than \emph{ppp}. 1km means 1 km aggregation of the data. The description
WorldPop gives for this data is «individual countries 2000-2020 UN adjusted
aggregated to 1km resolution using 100m resolution population count
datasets». See: https://hub.worldpop.org/doi/10.5258/SOTON/WP00671.
}
\examples{
downloadWorldPopData("COD", here("data", "geotiff"))
}
\author{
Bryce Carson
}
