% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{forecastError.cov}
\alias{forecastError.cov}
\title{Create a Forecast Error Covariance Matrix}
\usage{
forecastError.cov(
  layers,
  variableCovarianceFunction,
  forecastError.cov.sdBackground,
  forecastError.cor.length,
  neighbourhood.Bayes,
  compartmentsReported = 2
)
}
\arguments{
\item{layers}{The SpatRaster object with SVEIRD compartment layers, and a
layer classifying habitation. Created with the getSVEIRD.SpatRaster
function.}

\item{variableCovarianceFunction}{a covariance function used to determine
the covariance between two variables.}

\item{forecastError.cov.sdBackground}{The "background" standard deviation of
the covariances of the forecast covariance error matrix.}

\item{forecastError.cor.length}{"correlation length (i.e. the average size
of the fluctuations)," as stated by J. Murray on the Physics
StackExchange, \url{https://physics.stackexchange.com/a/671317}.}

\item{neighbourhood.Bayes}{TODO}

\item{compartmentsReported}{TODO}
}
\value{
TODO
}
\description{
generates a block diagonal error covariance matrix with exponential decay
}
\examples{
subregionsSpatVector <- terra::vect(
  system.file(
    "extdata",
    ## COD: Nord-Kivu and Ituri (Democratic Republic of Congo)
    "subregionsSpatVector",
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
susceptibleSpatRaster <- terra::rast(
  system.file(
    "extdata",
    "susceptibleSpatRaster.tif", # Congo population
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
layers <- getSVEIRD.SpatRaster(subregionsSpatVector,
                               susceptibleSpatRaster,
                               aggregationFactor = 10)
Ituri.forecastError.cov <- forecastError.cov(layers,
                                             variableCovarianceFunction = "DBD",
                                             forecastError.cov.sdBackground = 2,
                                             forecastError.cor.length = 0.8,
                                             neighbourhood = 1,
                                             compartmentsReported = 2)
}
\author{
Bryce Carson

Ashok Krishnmaurthy

Michael Myer

Thomas White
}
\keyword{internal}
