% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{getCountryPopulation.SpatRaster}
\alias{getCountryPopulation.SpatRaster}
\title{Retrieve a population count SpatRaster}
\usage{
getCountryPopulation.SpatRaster(countryCodeISO3C, folder = NULL)
}
\arguments{
\item{countryCodeISO3C}{The uppercase ISO three character code a recognized
country.}

\item{folder}{Passed on to method downloadWorldPopData}
}
\value{
A SpatRaster of WorldPop population count data with the layer name
"Population", with all NAs replaced by zeros.
}
\description{
Retrieve a SpatRaster useful for spatiotemporal epidemic compartmental
modelling.
}
\examples{
getCountryPopulation.SpatRaster("COD")
}
\author{
Bryce Carson

Michael Myer

Ashok Krishnmaurthy
}
