% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{getCountrySubregions.SpatVector}
\alias{getCountrySubregions.SpatVector}
\title{Get a country's GADM SpatVector, from GADM verison 3.6}
\usage{
getCountrySubregions.SpatVector(countryCodeISO3C = "COD", level1Region, folder)
}
\arguments{
\item{countryCodeISO3C}{The uppercase ISO three character code of a
recognized country.}

\item{level1Region}{Subregions of the country to limit the extent of the
SpatVector to.}

\item{folder}{The folder which should be searched for the GADM data, passed
on to lvl1AdminBorders.}
}
\value{
SpatVector for the specificed country, optionally cropped to the
provinces/states specified in the \code{level1Region} argument.
}
\description{
Acquire a SpatVector for a given country, optionally limited to subregion(s) thereof.
}
\examples{
library(geodata)
geodata_path_backup <- geodata_path()
options(geodata_default_path = getOption("spatialEpisim.foundation.datapath"))
getCountrySubregions.SpatVector("COD")
getCountrySubregions.SpatVector("COD", c("Nord-Kivu", "Ituri"))
getCountrySubregions.SpatVector("CZE", "Prague")
getCountrySubregions.SpatVector("NGA", "Kwara")
options(geodata_default_path = geodata_path_backup)
}
\author{
Bryce Carson
}
