% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{healthZonesCongo}
\alias{healthZonesCongo}
\title{Reporting health zones of the Democratic Republic of Congo}
\format{
\subsection{\code{healthZonesCongo}}{

\describe{
\item{HealthZone}{The name of the Congolese health zone}
\item{Latitude}{The geographical latitude of the health zone}
\item{Longitude}{The geographical longitude of the health zone}
}

\preformatted{
       HealthZone   Latitude Longitude
  1     Alimbongo -0.3655150  29.19118
  2          Beni  0.4911300  29.47306
  3         Biena  0.5792300  29.11563
  4       Butembo  0.1406920  29.33501
  5          Goma -1.6582710  29.22013
}
}
}
\usage{
healthZonesCongo
}
\description{
Reporting health zones of the Democratic Republic of Congo
}
\keyword{datasets}
