% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{initialInfections.fourCities}
\alias{initialInfections.fourCities}
\title{2018 Congo Ebola epidmeic data}
\format{
\subsection{\code{initialInfections.fourCities}}{

\describe{
\item{Location}{The name of the location (reporting health zone) which the other variables describe}
\item{Latitude}{The geographical latitude of the health zone}
\item{Longitude}{The geographical longitude of the health zone}
\item{InitialVaccinated}{The number of vaccinated people}
\item{InitialExposed}{The number of exposed people}
\item{InitialInfections}{The number of infected people}
\item{InitialRecovered}{The number of recovered people}
\item{InitialDead}{The number of dead people}
}
}
}
\usage{
initialInfections.fourCities
}
\description{
"Seed" data to use for a simulation of the 2018 Ebola epidemic in the Congo
}
\details{
Here is an example the data used.

\preformatted{
  Location Latitude Longitude Vaccinated Exposed Infected Recovered Dead
  Beni     0.49113  29.47306  0          24      12       0         4
  Butembo  0.140692 29.335014 0          0       0        0         0
  Mabalako 0.461257 29.210687 0          0       0        0         0
  Mandima  1.35551  29.08173  0          0       0        0         0
}
}
\keyword{datasets}
