% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{linearInterpolationOperator}
\alias{linearInterpolationOperator}
\title{Linear (Forward) Interpolation Operator matrices for one or two state
vectors}
\usage{
linearInterpolationOperator(
  layers,
  healthZoneCoordinates,
  neighbourhood.order = 2,
  compartmentsReported = 1
)
}
\arguments{
\item{layers}{The SpatRaster object with SVEIRD compartment layers, and a
layer classifying habitation. Created with the getSVEIRD.SpatRaster
function.}

\item{healthZoneCoordinates}{a table of values giving the latitude and
longitude coordinates for health zones in the country of interest. See the
description of the healthZoneCoordinates argument in the function
SVEIRD.BayesianDataAssimilation for more details.

\preformatted{
    HealthZone    Latitude    Longitude
    Alimbongo     -0.365515   29.1911818
    Beni          0.49113     29.47306
    Biena         0.57923     29.115633
    Butembo       0.140692    29.335014
    Goma          -1.658271   29.220132
    Kalunguta     0.323085    29.354677
    Katwa         0.116985    29.411838
    Kayna         -0.603936   29.174066
    Kyondo        -0.005622   29.408813
    Lubero        -0.15633    29.24057
    Mabalako      0.461257    29.210687
    Manguredjipa  0.353433    28.765479
    Masereka      -0.133333   29.333333
    Musienene     0.04022     29.26246
    Mutwanga      0.32893     29.74576
    Nyiragongo    -1.516667   29.25
    Oicha         0.698681    29.518834
    Pinga         -0.9830504  28.687911
    Vuhovi        0.1416      29.4075
    Ariwara       3.136873    30.706615
    Bunia         1.566667    30.25
    Komanda       1.367496    29.774322
    Lolwa         1.352969    29.498455
    Mambasa       1.359731    29.029226
    Mandima       1.35551     29.08173
    Nyakunde      1.431271    30.029283
    Rwampara      1.4053      30.3449
    Tchomia       1.4412      30.4845
  }}

\item{compartmentsReported}{either 1 or 2. Previously identified as
states_observable, this is the count of compartments that are reported on
and which will have data assimilated; if it is 2, the matrix is a block
diagonal matrix.}
}
\value{
a linear forward interpolation operator matrix used in
SVEIRD.BayesianDataAssimilation simulations.
}
\description{
Linear (Forward) Interpolation Operator matrices for one or two state
vectors
}
\details{
Create a linear forward interpolation operator matrix with as many
columns as cells in the SpatRaster layers, and as many rows as health
zones for which coordinates are provided (when creating a matrix for use
with one state vector). The matrix is either a trivial matrix as
described, or two partitions in a sparse, block diagonal matrix.

When used to creat an interpolation matrix for two state vectors, the result
is a block diagonal matrix of identical partitions, with each partition as
described for one state vector.

The dimensions of the returned matrix are nrow(healthZoneCoordinates) rows
by ncell(layers) columns.
}
\examples{
subregionsSpatVector <- terra::vect(
  system.file(
    "extdata",
    ## COD: Nord-Kivu and Ituri (Democratic Republic of Congo)
    "subregionsSpatVector",
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
susceptibleSpatRaster <- terra::rast(
  system.file(
    "extdata",
    "susceptibleSpatRaster.tif", # Congo population
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
data("healthZonesCongo", package = "spatialEpisim.foundation")
linearInterpolationOperator(
  layers = getSVEIRD.SpatRaster(subregionsSpatVector,
                                susceptibleSpatRaster,
                                aggregationFactor = 10),
  healthZoneCoordinates = healthZonesCongo
)

H.matrix <-
  linearInterpolationOperator(
    layers = getSVEIRD.SpatRaster(subregionsSpatVector,
                                  susceptibleSpatRaster,
                                  aggregationFactor = 10),
    healthZoneCoordinates = healthZonesCongo,
    neighbourhood.order = 1
  )
plot(as.array(H.matrix[1, ]))

linearInterpolationOperator(
  layers = getSVEIRD.SpatRaster(subregionsSpatVector,
                                susceptibleSpatRaster,
                                aggregationFactor = 10),
  healthZoneCoordinates = healthZonesCongo,
  neighbourhood.order = 2
)
}
\author{
Bryce Carson

Ashok Krishnmaurthy

Michael Myer
}
\keyword{internal}
