% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{lvl1AdminBorders}
\alias{lvl1AdminBorders}
\title{Retrieve a SpatRaster of level 1 boundaries from local or remote disk}
\usage{
lvl1AdminBorders(countryISO3C, folder = geodata_path())
}
\arguments{
\item{countryISO3C}{The uppercase ISO three character code a recognized
country.}

\item{folder}{The path where GADM data should be found or stored; passed on
to geodata::gadma.}
}
\value{
SpatVector
}
\description{
Returns a SpatVector for the requested country on demand, either retrieving
the data from disk if it has been downloaded before, or downloading it for
the first time.
}
\details{
The level one boundaries are the least granular administrative boundaries
that countries create to subdivide themselves.
}
\examples{
geodata_path_backup <- geodata_path()
options(geodata_default_path = tempdir())
lvl1AdminBorders("COD")
options(geodata_default_path = geodata_path_backup)

\dontrun{
## It's recommended to set geodata_default_path rather than explicitly set
## the path
lvl1AdminBorders("COD", file.path("data", "gadm"))
}
}
\author{
Bryce Carson
}
