% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{maskAndClassifySusceptibleSpatRaster}
\alias{maskAndClassifySusceptibleSpatRaster}
\title{Plot a raster cropped to sub-region(s) of a country}
\usage{
maskAndClassifySusceptibleSpatRaster(subregions, susceptible)
}
\arguments{
\item{subregions}{a SpatRaster object for subregions, created with
getCountrySubregions.SpatVector}

\item{susceptible}{SpatRaster or SpatVector to be masked by GADM data
of the country}
}
\description{
Plot a raster cropped to a specific region(s)
}
\details{
As a side effect, it writes a raster to file.
}
\examples{
subregionsSpatVector <- terra::vect(
  system.file(
    "extdata",
    ## COD: Nord-Kivu and Ituri (Democratic Republic of Congo)
    "subregionsSpatVector",
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
susceptibleSpatRaster <- terra::rast(
  system.file(
    "extdata",
    "susceptibleSpatRaster.tif", # Congo population
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
levelplot(maskAndClassifySusceptibleSpatRaster(subregionsSpatVector,
                                               susceptibleSpatRaster))

\dontrun{
maskAndClassifySusceptibleSpatRaster(getSubregion("CZE", "Prague"),
                                     getCountryPopulation.SpatRaster("CZE"))

maskAndClassifySusceptibleSpatRaster(getSubregion("NGA", "Lagos"),
                                     getCountryPopulation.SpatRaster("NGA"))

maskAndClassifySusceptibleSpatRaster(getSubregion("COD", "Ituri"),
                                     getCountryPopulation.SpatRaster("COD"))

maskAndClassifySusceptibleSpatRaster(getSubregion("COD", c("Nord-Kivu", "Ituri")),
                                     getCountryPopulation.SpatRaster("COD"))
}
}
\author{
Bryce Carson

gursDhaliwal

Ashok Krishnmaurthy

Michael Myer

Thomas White
}
\keyword{internal}
