% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{Q.forecastErrorCov}
\alias{Q.forecastErrorCov}
\title{Create a Q-matrix}
\usage{
Q.forecastErrorCov(
  layers,
  variableCovarianceFunction,
  Q.backgroundErrorStandardDeviation,
  Q.characteristicCorrelationLength,
  neighbourhood,
  compartmentsReported = 2
)
}
\arguments{
\item{layers}{The SpatRaster object with SVEIRD compartment layers, and a
layer classifying habitation. Created with the getSVEIRD.SpatRaster
function.}

\item{variableCovarianceFunction}{TODO}

\item{Q.characteristicCorrelationLength}{TODO}

\item{neighbourhood}{TODO}

\item{compartmentsReported}{TODO}

\item{Q.backgroundErrorStandardDeviatio}{TODO}
}
\value{
TODO
}
\description{
generates a block diagonal error covariance matrix with exponential decay
}
\details{
TODO: write the details about the implementation of this function.
}
\examples{
subregionsSpatVector <- terra::vect(
  system.file(
    "extdata",
    ## COD: Nord-Kivu and Ituri (Democratic Republic of Congo)
    "subregionsSpatVector",
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
susceptibleSpatRaster <- terra::rast(
  system.file(
    "extdata",
    "susceptibleSpatRaster.tif", # Congo population
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
layers <- getSVEIRD.SpatRaster(subregionsSpatVector,
                               susceptibleSpatRaster,
                               aggregationFactor = 35)
Ituri.Q.forecastErrorCov <- Q.forecastErrorCov(layers, "DBD", 2, 0.8, 4, 2)
}
\author{
Bryce Carson

Ashok Krishnmaurthy

Michael Myer

Thomas White
}
\keyword{internal}
