% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{reclassifyBelowUpperBound}
\alias{reclassifyBelowUpperBound}
\title{Reclassify Below Upper Bound}
\usage{
reclassifyBelowUpperBound(raster, upper, reclassification = 0)
}
\arguments{
\item{raster}{a SpatRaster to reclassify}

\item{upper}{the upper bound, not included in the range of values classified}

\item{reclassification}{the value to use when reclassifying those below the upper value (zero by default).}
}
\value{
the reclassified SpatRaster
}
\description{
Change all numbers in a raster which are less than some upper value to zero.
}
\details{
With an upper of zero, negatives will be reclassified to zero, whilst with
an upper bound of some positive number, any number less than that will be
reclassified to zero.
}
\author{
Bryce Carson
}
