% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{replaceInequalityWith}
\alias{replaceInequalityWith}
\title{Replace logical vectors derived from inequalities with z value}
\usage{
replaceInequalityWith(f, w, x, y = NULL, z)
}
\arguments{
\item{f}{the binary operator function to use to calculate the logical vector
for subsetting w.}

\item{w}{the vector to subset and subassign within.}

\item{x}{the first argument of f.}

\item{y}{the second argument of y (if applicable).}

\item{z}{the value to assign to the logical subset of w.}
}
\value{
w, with the modified values, so that it can be used in a pipe.
}
\description{
This is a convenience function to make subassignment a little easier. It
doesn't need names like subset appears to need, and it allows passing
functions as values.
}
\details{
The value z is used to replace the vector elements with the provided value
z (zed).

The vector subset, w, is subset with the function f of x and x, and z
assigned to these elements. See the function body for more details.
}
\author{
Bryce Carson
}
\keyword{internal}
