% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{setupBayesianDataAssimilation}
\alias{setupBayesianDataAssimilation}
\title{Setup Bayesian data assimilation}
\usage{
setupBayesianDataAssimilation(
  layers,
  healthZoneCoordinates,
  compartmentsReported,
  variableCovarianceFunction,
  Q.backgroundErrorStandardDeviation,
  Q.characteristicCorrelationLength,
  neighbourhood
)
}
\arguments{
\item{layers}{a SpatRaster with the following layers: Susceptible,
Vaccinated, Exposed, Infected, Recovered, and Dead.}

\item{healthZoneCoordinates}{a dataframe with three columns, a location
name, latitude, and longitude describing the geographical locations of
reporting health zones.}

\item{compartmentsReported}{the number of compartments or epidemic states
reported on; either one or two. Higher numbers are not supported. One
corresponds only to infection, while two compartments corresponds to
exposure and infection.}

\item{variableCovarianceFunction}{a function to calculate the error covariance, returned by Q.forecastErrorCov.}

\item{Q.backgroundErrorStandardDeviation}{the "background" or default amount of error, in standard deviations.}

\item{Q.characteristicCorrelationLength}{TODO}

\item{neighbourhood}{TODO}
}
\value{
the HQHt matrix
}
\description{
Using the provided parameters and SpatRaster, the necessary setup functions
and values for Bayesian data assimilation are called, with values used later
on returned.
}
\author{
Bryce Carson
}
