% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialEpisim.foundation-package.R
\docType{package}
\name{spatialEpisim.foundation-package}
\alias{spatialEpisim.foundation}
\alias{spatialEpisim.foundation-package}
\title{spatialEpisim.foundation: Spatial, Compartmental Epidemic Simulations with Exposure, Vaccination, and Bayesian Data Assimilation for 'spatialEpisim'}
\description{
Spatial epidemic simulations using SVEIRD compartmental epidemic models (Susceptible, Vaccinated, Exposed, Infected, Recovered, Dead) allow the mathematical modelling of epidemics using traditional, modified SIR (Susceptible, Infected, Recovered) compartmental models as invented by Kermack and McKendrick in 1927. Spatial dynamics extend the SVEIRD modelling process in this software. 'spatialEpisim' is a 'Shiny' application built on this 'spatialEpisim.foundation'. The sources originated there, and now live here to provide for further validation and documenation through vignettes, tests, et cetera.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krishnamurthylab/spatialepisim.foundation}
  \item \url{https://krishnamurthylab.github.io/spatialepisim.foundation/}
  \item Report bugs at \url{https://github.com/krishnamurthylab/spatialepisim.foundation/issues}
}

}
\author{
\strong{Maintainer}: Ashok Krishnamurthy \email{akrishnamurthy@mtroyal.ca} (\href{https://orcid.org/0000-0003-1517-070X}{ORCID}) [copyright holder, contributor]

Authors:
\itemize{
  \item Bryce Carson \email{bcars268@mtroyal.ca} (\href{https://orcid.org/0000-0002-1362-2998}{ORCID})
}

Other contributors:
\itemize{
  \item Alberta Innovates [funder]
}

}
\keyword{internal}
