% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{SVEIRD.BayesianDataAssimilation}
\alias{SVEIRD.BayesianDataAssimilation}
\title{SVEIRD compartmental model with optional Bayesian data assimilation}
\usage{
SVEIRD.BayesianDataAssimilation(
  alpha,
  beta,
  gamma,
  sigma,
  delta,
  radius,
  lambda,
  n.days,
  seedData,
  seedRadius = 0,
  layers,
  rasterAgg,
  startDate,
  countryISO3C,
  incidenceData = NULL,
  deathData = NULL,
  simulationIsDeterministic = TRUE,
  dataAssimilationEnabled = FALSE,
  healthZoneCoordinates,
  variableCovarianceFunction,
  Q.backgroundErrorStandardDeviation,
  Q.characteristicCorrelationLength,
  neighbourhood,
  psi.diagonal,
  callback = `{`
)
}
\arguments{
\item{alpha}{The rate of vaccination (per day)}

\item{beta}{The rate of exposure (per day)}

\item{gamma}{The rate of becoming infectious (per day)}

\item{sigma}{The rate of recovery (per day)}

\item{delta}{The fatality rate (per day)}

\item{radius}{The distance, in kilometers, a given individual travels from
their starting point (on average, per day)}

\item{lambda}{The probability that an individual will move the distance
governed by radius}

\item{n.days}{The number of days the simulation will run for, beginning from
the startDate.}

\item{seedData}{a dataframe like the following example; the compartment
columns are the initial values.

\preformatted{ Location Latitude Longitude Vaccinated Exposed Infected
    Recovered Dead Beni 0.49113 29.47306 0 24 12 0 4 Butembo 0.140692
    29.335014 0 0 0 0 0 Mabalako 0.461257 29.210687 0 0 0 0 0 Mandima
    1.35551 29.08173 0 0 0 0 0 }}

\item{seedRadius}{The number of cells over which to average the seed data in
a Moore neighbourhood for each locality.}

\item{layers}{The SpatRaster object with SVEIRD compartment layers, and a
layer classifying habitation. Created with the getSVEIRD.SpatRaster
function.}

\item{rasterAgg}{The number of adjacent cells in any one direction to
aggregate into a single cell. The aggregation factor must be the same as
that used to generate the SpatRaster for layers.}

\item{startDate}{The date (in YYYY-MM-DD format) the simulation begins.}

\item{countryISO3C}{The ISO three character code for a recognized country.}

\item{incidenceData}{A "situation report" dataframe. The first column
provides the date of the officially reported, observed incidence of the
disease, in ISO format (YYYY-MM-DD). MAYBE TODO: enforce the startDate
parameter to be one week prior to the first observed data?

\preformatted{ Date Beni Butembo Mabalako Mandima 2018-08-05 34 34 34 34
    2018-08-12 2 0 11 1 2018-08-20 1 0 37 6 2018-08-26 5 0 3 0 2018-08-02 8
    0 1 1 2018-08-09 5 2 1 1 }}

\item{deathData}{Data of the same format as incidenceData, but observations
represent deaths, not infections.}

\item{simulationIsDeterministic}{Whether stochasticity is enabled or not; if
the simulation is deterministic then no stochastic processes are used and
the simulation is entirely deterministic.}

\item{dataAssimilationEnabled}{Whether Bayesian data assimilation will be
used for state reporting data.}

\item{healthZoneCoordinates}{The coordinates of health zones in the country
of interest, TODO: describe the use of the data so users understand why
they must include it; e.g. (which will be used to group and summarize the
compartmental model data at the end of the simmulation.)

\preformatted{
    HealthZone    Latitude    Longitude
    Alimbongo     -0.365515   29.1911818
    Beni          0.49113     29.47306
    Biena         0.57923     29.115633
    Butembo       0.140692    29.335014
    Goma          -1.658271   29.220132
    Kalunguta     0.323085    29.354677
    Katwa         0.116985    29.411838
    Kayna         -0.603936   29.174066
    Kyondo        -0.005622   29.408813
    Lubero        -0.15633    29.24057
    Mabalako      0.461257    29.210687
    Manguredjipa  0.353433    28.765479
    Masereka      -0.133333   29.333333
    Musienene     0.04022     29.26246
    Mutwanga      0.32893     29.74576
    Nyiragongo    -1.516667   29.25
    Oicha         0.698681    29.518834
    Pinga         -0.9830504  28.687911
    Vuhovi        0.1416      29.4075
    Ariwara       3.136873    30.706615
    Bunia         1.566667    30.25
    Komanda       1.367496    29.774322
    Lolwa         1.352969    29.498455
    Mambasa       1.359731    29.029226
    Mandima       1.35551     29.08173
    Nyakunde      1.431271    30.029283
    Rwampara      1.4053      30.3449
    Tchomia       1.4412      30.4845
  }}

\item{variableCovarianceFunction}{Passed directly to \code{\link[=Q.forecastErrorCov]{Q.forecastErrorCov()}} to generate a
Q matrix.}

\item{Q.backgroundErrorStandardDeviation}{TODO}

\item{Q.characteristicCorrelationLength}{TODO}

\item{neighbourhood}{TODO}

\item{psi.diagonal}{TODO}

\item{callback}{a callback function to run, with no arguments, which will be
called every time the main loop of the simulation iterates.}
}
\value{
a summary dataframe for the simulation, showing changes in the
compartment values over time, the daily values, and cumulative values.
}
\description{
Run a SVEIRD compartmental model of an epidemic, optionally
using Bayesian data assimilation.
}
\details{
TODO: DETAILS of the function.
}
\examples{
subregionsSpatVector <- terra::vect(
  system.file(
    "extdata",
    ## COD: Nord-Kivu and Ituri (Democratic Republic of Congo)
    "subregionsSpatVector",
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
susceptibleSpatRaster <- terra::rast(
  system.file(
    "extdata",
    "susceptibleSpatRaster.tif", # Congo population
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
SVEIRD.BayesianDataAssimilation(
  ## Parameters
  alpha = 3.5e-5,
  beta = 7e-3,
  gamma = 1/7,
  sigma = 1/36,
  delta = 2/36,
  lambda = 15,
  ## Model runtime
  n.days = 31, # a month, permitting three assimilations of observed data
  ## Model data
  seedData = data("initialInfections.fourCities"),
  seedRadius = 1,
  layers = getSVEIRD.SpatRaster(subregionsSpatVector,
                                susceptibleSpatRaster,
                                aggregationFactor = 35),
  rasterAgg = 35,
  startDate = "2018-08-05",
  countryISO3C = "COD",
  incidenceData = data("Congo.EbolaIncidence"),
  ## Model options
  dataAssimilationEnabled = TRUE,
  healthZoneCoordinates = data("healthZonesCongo"),
  variableCovarianceFunction = "DBD",
  ## Special parameters
  Q.backgroundErrorStandardDeviation = 0.55,
  Q.characteristicCorrelationLength = 6.75e-1,
  neighbourhood = 3,
  psi.diagonal = 1e-3
)
}
\author{
Bryce Carson

Ashok Krishnmaurthy

Michael Myer
}
