% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{transmissionLikelihoodWeightings}
\alias{transmissionLikelihoodWeightings}
\title{Weighted Sums}
\usage{
transmissionLikelihoodWeightings(infections, lambda, aggregationFactor)
}
\arguments{
\item{infections}{a matrix of the count of infections per aggregate area in
a raster of terrestrial data.}

\item{lambda}{the average dialy movement distance of an individual (in
kilometers).}

\item{aggregationFactor}{The number of adjacent cells in any one direction to
aggregate into a single cell. The aggregation factor must be the same as
that used to generate the SpatRaster for layers.}
}
\value{
a matrix of weightings for the calculation of the proportionOfSusceptible of
exposed individuals who will become infectious.
}
\description{
Calculate a matrix of weights respecting human mobility
patterns.
}
\details{
The pattern of human mobility used is described in a slideshow
here:
https://docs.google.com/presentation/d/1_gqcEh4d8yRy22tCZkU0MbGYSsGu3Djh/edit?usp=sharing&ouid=102231457806738400087&rtpof=true&sd=true.
}
\examples{
library(terra)
subregionsSpatVector <- vect(
  system.file(
    "extdata",
    ## COD: Nord-Kivu and Ituri (Democratic Republic of Congo)
    "subregionsSpatVector",
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
susceptibleSpatRaster <- rast(
  system.file(
    "extdata",
    "susceptibleSpatRaster.tif", # Congo population
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
layers <- getSVEIRD.SpatRaster(subregionsSpatVector,
                               susceptibleSpatRaster,
                               aggregationFactor = 35)
transmissionLikelihoodWeightings(layers$Infected, 15, 35)
}
\author{
Bryce Carson

Thomas White
}
\keyword{internal}
