% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{getCountrySubregions.SpatVector}
\alias{getCountrySubregions.SpatVector}
\title{Get a country's GADM verison 3.6 raster}
\usage{
getCountrySubregions.SpatVector(
  countryISO3C = "COD",
  level1Region = c("Nord-Kivu", "Ituri"),
  folder
)
}
\arguments{
\item{countryISO3C}{The uppercase ISO three character code a recognized
country.}

\item{level1Region}{The subregions of the country to crop to}

\item{folder}{The folder which should be searched for the GADM data, passed
on to lvl1AdminBorders.}
}
\value{
SpatVector for the specificed country.
}
\description{
Read an RDS file from the GADM folder for a given country's subregion.
}
\examples{
geodata_path_backup <- geodata_path()
options(geodata_default_path = tempdir())
getCountrySubregions.SpatVector("COD", c("Nord-Kivu", "Ituri"))
getCountrySubregions.SpatVector("CZE", "Prague")
getCountrySubregions.SpatVector("NGA", "Kwara")
options(geodata_default_path = geodata_path_backup)
}
\author{
Bryce Carson
}
