% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{getSVEIRD.SpatRaster}
\alias{getSVEIRD.SpatRaster}
\title{Create a SpatRaster of SVEIRD model compartment raster data}
\usage{
getSVEIRD.SpatRaster(subregions, susceptible, aggregationFactor = NULL)
}
\arguments{
\item{subregions}{a SpatVector object of subregions used to crop the SVEIRD
SpatRaster}

\item{susceptible}{a RasterLayer, pre-aggregated if that is wished.}

\item{aggregationFactor}{the number of cells in any direction to aggregate
together into one, in the susceptible ratser, after masking with the
subregions vector, before creating the list}
}
\value{
a SpatRaster, with layers for the SVERID components
}
\description{
Create a list of SpatRaster objects, one for each component in
an SVEIRD epidemic model.
}
\details{
The SpatRaster objects for the VEIRD components are empty, the
input SpatRaster is the only layer with non-zero values (if any existed before).
}
\examples{
subregionsSpatVector <- terra::vect(
  system.file(
    "extdata",
    ## COD: Nord-Kivu and Ituri (Democratic Republic of Congo)
    "subregionsSpatVector",
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
susceptibleSpatRaster <- terra::rast(
  system.file(
    "extdata",
    "susceptibleSpatRaster.tif", # Congo population
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
getSVEIRD.SpatRaster(subregionsSpatVector, susceptibleSpatRaster,
                     aggregationFactor = 10)

## Omitting the aggregation factor argument will prevent aggregation. An
## aggregation factor of zero or one is meaningless and will produce an
## error.
getSVEIRD.SpatRaster(subregionsSpatVector, susceptibleSpatRaster)
}
\author{
Bryce Carson

Michael Myer

Ashok Krishnmaurthy
}
