% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{getSVEIRD.SpatRaster}
\alias{getSVEIRD.SpatRaster}
\title{Create a SpatRaster of SVEIRD model compartment raster data}
\usage{
getSVEIRD.SpatRaster(subregions, population, aggregationFactor = NULL)
}
\arguments{
\item{subregions}{a SpatVector object of subregions used to crop the
population SpatRaster before creating the other layers in the raster
object to represent the other compartments in a SVEIRD epidemic model. If
it is missing no cropping is performed.}

\item{population}{a SpatRaster of population count data; it must be
pre-aggregated if any aggregation is to be used during the simulation;
none is performed by this function or downstream functions in the overall
implementation of \link{SVEIRD.BayesianDataAssimilation} simulations.}

\item{aggregationFactor}{the number of cells in any direction to aggregate
together into one, in the input raster, after masking with the subregions
vector.}
}
\value{
a SpatRaster, with layers for the SVERID components
}
\description{
Create a SpatRaster objects, with a layer for each component in
an SVEIRD epidemic model.
}
\details{
The SpatRaster objects for the VEIRD components are empty, the
input SpatRaster is taken as the Suscptible layer, the only layer with
non-zero values (if any existed before).
}
\examples{
subregionsSpatVector <- terra::vect(
  system.file(
    "extdata",
    ## COD: Nord-Kivu and Ituri (Democratic Republic of Congo)
    "subregionsSpatVector",
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
susceptibleSpatRaster <- terra::rast(
  system.file(
    "extdata",
    "susceptibleSpatRaster.tif", # Congo population
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
getSVEIRD.SpatRaster(subregionsSpatVector, susceptibleSpatRaster,
                     aggregationFactor = 10)

## Omitting the aggregation factor argument will prevent aggregation. An
## aggregation factor of zero or one is meaningless and will produce an
## error.
getSVEIRD.SpatRaster(subregionsSpatVector, susceptibleSpatRaster)

## Omitting a SpatVector will return a SpatRaster with layers representing each
## of SVEIRD without cropping or masking the input SpatRaster in any way.
getSVEIRD.SpatRaster(subregionsSpatVector, susceptibleSpatRaster)
}
\author{
Bryce Carson

Michael Myer

Ashok Krishnmaurthy
}
