% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{initialInfections.fourCities}
\alias{initialInfections.fourCities}
\title{Seed data to use for a simulation of the 2018 Ebola epidmeic in the Congo}
\format{
\subsection{\code{initialInfections.fourCities}}{

\describe{
\item{Location}{The name of the location (reporting health zone) which the other variables describe}
\item{Latitude}{The geographical latitude of the health zone}
\item{Longitude}{The geographical longitude of the health zone}
\item{InitialVaccinated}{The number of vaccinated people}
\item{InitialExposed}{The number of exposed people}
\item{InitialInfections}{The number of infected people}
\item{InitialRecovered}{The number of recovered people}
\item{InitialDead}{The number of dead people}
}
}
}
\usage{
initialInfections.fourCities
}
\description{
Seed data to use for a simulation of the 2018 Ebola epidmeic in the Congo
}
\keyword{datasets}
