% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVEIRD.BayesianDataAssimilation.R
\name{transmissionLikelihoodWeightings}
\alias{transmissionLikelihoodWeightings}
\title{Weighted Sums}
\usage{
transmissionLikelihoodWeightings(infections, radius, lambda, aggregationFactor)
}
\arguments{
\item{infections}{a matrix of the count of infections per aggregate area in
a raster of terrestrial data.}

\item{radius}{a constant; see details.}

\item{lambda}{movemenet distance (in kilometers) per day; see details.}

\item{aggregationFactor}{the degree of aggregation applied to the raster
data mentioned in the function details.}
}
\value{
a matrix of weightings for the calculation of the proportion of
exposed individuals who will become infectious.
}
\description{
Calculate a matrix of weights respecting human mobility
patterns.
}
\details{
The pattern of human mobility used is described in a slideshow
here:
https://docs.google.com/presentation/d/1_gqcEh4d8yRy22tCZkU0MbGYSsGu3Djh/edit?usp=sharing&ouid=102231457806738400087&rtpof=true&sd=true.
}
\examples{
## TODO FIXME
terra::as.matrix(Infected, wide = TRUE) \%>\%
  transmissionLikelihoodWeightings(30, 15, 35)
}
\author{
Bryce Carson

Thomas White
}
\keyword{internal}
